requirements.txt
keyboard==0.13.5
pystray==0.19.4
Pillow>=10.0.0
